+++ mmodel/model.py
@@ -1,16 +1,11 @@
-from mmodel.utility import (
-    parse_input,
-    parse_modifiers,
-    content_wrap,
-    is_node_attr_defined,
-    is_edge_attr_defined,
-)
+from mmodel.utility import parse_input, is_node_attr_defined, is_edge_attr_defined
+from mmodel.metadata import modelformatter, textwrap80, format_metadata
 from mmodel.draw import draw_graph
 import networkx as nx
 
 
 class Model:
-    """Create model executable.
+    """Create model callable.
 
     :param str name: Model name
     :param object graph: ModelGraph instance (digraph)
@@ -26,6 +21,8 @@ class Model:
     :param list returns: The order of returns of the model; defaults to the
         topological search.
     """
+
+    _model_keys = ["name", "graph", "handler", "modifiers", "description", "returns"]
 
     def __init__(
         self,
@@ -50,6 +47,7 @@ class Model:
         handler_class, handler_kwargs = handler
 
         executor = handler_class(name, self._graph, self.returns, **handler_kwargs)
+        self.execution_order = [node for node, _ in executor.order]
 
         for mdf, kwargs in self.modifiers:
             executor = mdf(executor, **kwargs)
@@ -57,48 +55,48 @@ class Model:
         self.__signature__ = executor.__signature__
 
         # final callable
-        self.executor = executor
+        self._executor = executor
 
     def __call__(self, *args, **kwargs):
+        """Execute the model.
+
+        The inputs from the keyword arguments are parsed and passed to the
+        the handler class.
+        """
 
         # process inputs
         inputs = parse_input(self.__signature__, *args, **kwargs)
 
-        return self.executor(**inputs)
+        return self._executor(**inputs)
+
+    def _metadata_dict(self, verbose):
+        """Return a dictionary with metadata keys."""
+
+        short_dict = {"model": self, "returns": self.returns}
+
+        additonal_dict = {
+            "graph": self._graph,
+            "handler": self.handler,
+            "modifiers": self.modifiers,
+            "description": self.description,
+        }
+
+        if verbose:
+            return {**short_dict, **additonal_dict}
+        else:
+            return short_dict
+
+    def metadata_str(
+        self, verbose=True, formatter=modelformatter, textwrapper=textwrap80
+    ):
+        """Parse metadata string of the Model instance."""
+        return format_metadata(self._metadata_dict(verbose), formatter, textwrapper)
 
     def __str__(self):
         """Output callable information."""
 
-        return self.metadata()
-
-    def metadata(self, full=True, wrap_width=80):
-        """Parse metadata string of the Model instance."""
-
-        # use tuple if there are multiple returns
-        # else use returns directly.
-        return_len = len(self.returns)
-        if return_len == 0:
-            returns_str = "None"
-        elif return_len == 1:
-            returns_str = self.returns[0]
-        else:
-            returns_str = f"({', '.join(self.returns)})"
-
-        metadata_list = [
-            f"{self.__name__}{self.__signature__}",
-            f"returns: {returns_str}",
-            f"handler: {self.handler[0].__name__}"
-            f"({', '.join(repr(v) for v in self.handler[1].values())})",
-        ]
-
-        metadata_list.extend(parse_modifiers(self.modifiers))
-
-        if full:
-            metadata_list.extend(["", self.description])
-
-        wrapped_list = content_wrap(metadata_list, width=wrap_width)
-
-        return "\n".join(wrapped_list).rstrip()
+        str_list = self.metadata_str()
+        return "\n".join(str_list).rstrip()
 
     @staticmethod
     def _is_valid_graph(G):
@@ -135,19 +133,19 @@ class Model:
 
     def get_node_func(self, node):
         """Quick access to node base callable within the model.
-        
+
         The function helps extract the original function within
         the node.
         """
 
         return self._graph.nodes[node]["_func"]
 
-    def node_metadata(self, node, full=True, wrap_width=80):
+    def node_metadata(self, *args, **kwargs):
         """View a specific node."""
 
-        return self._graph.node_metadata(node, full=True, wrap_width=80)
+        return self._graph.node_metadata(*args, **kwargs)
 
-    def draw(self, style="full", export=None, wrap_width=30):
+    def draw(self, style="verbose", export=None):
         """Draw the graph of the model.
 
         Draws the default styled graph.
@@ -157,7 +155,17 @@ class Model:
             long shows all the metadata.
         :param str export: filename to save the graph as. The file extension
             is needed.
-
         """
 
-        return draw_graph(self._graph, self.metadata(), style, export, wrap_width)
+        return draw_graph(self._graph, str(self), style, export)
+
+    def edit(self, **kwargs):
+        """Edit components of the model to create a new model.
+
+        It is not recommended to edit the graph component of the model.
+        Although it does create a new model, but "edit" is for creating
+        a new model with the same graph.
+        """
+
+        model_dict = {key: getattr(self, key) for key in self._model_keys}
+        return self.__class__(**{**model_dict, **kwargs})